Added package to the application.
