+++ flaskblog/requirements.txt
